application/datamanager/pyproject.toml
[project]
name = "datamanager"
version = "0.1.0"
description = "Data management service"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "duckdb>=1.2.2",
    "polars>=1.29.0",
    "pyarrow>=20.0.0",
    "datamanager",
]

[tool.hatch.build.targets.wheel]
packages = ["datamanager"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest-asyncio>=0.26.0",
    "behave>=1.2.6",
    "requests>=2.31.0",
]
